Memory allocators: Revert Q_MALLOC to 8-byte alignment
Commit 1f55d05c2b actually upped Q_MALLOC's address alignment from
8-bytes to 16-bytes.  While this isn't necessarily a bug (in fact,
it makes OpenSIPS more crash-safe), it enables new scenarios where
buffer overrun crashes would happen with F_MALLOC but not with Q_MALLOC.

And since Q_MALLOC is the intended crash debugger, these scenarios
must not be allowed, so we stick to 8-byte alignment.  If anyone needs
16-byte aligned memory, they can just re-define ROUNDTO and re-compile.

(cherry picked from commit c29652fa10dbd34ce499f817891abe25964ee63b)

+++ mem/hp_malloc.c
@@ -527,7 +527,7 @@ static struct hp_block *hp_malloc_init(char *address, unsigned long size,
 	/* make address and size multiple of 8*/
 	start = (char *)ROUNDUP((unsigned long) address);
 	LM_DBG("HP_OPTIMIZE=%lu, HP_LINEAR_HASH_SIZE=%lu, %lu-bytes aligned\n",
-			HP_MALLOC_OPTIMIZE, HP_LINEAR_HASH_SIZE, ROUNDTO);
+			HP_MALLOC_OPTIMIZE, HP_LINEAR_HASH_SIZE, (unsigned long)ROUNDTO);
 	LM_DBG("HP_HASH_SIZE=%lu, HP_EXTRA_HASH_SIZE=%lu, hp_block size=%zu, "
 			"frag_size=%zu\n", HP_HASH_SIZE, HP_EXTRA_HASH_SIZE,
 			sizeof(struct hp_block), sizeof(struct hp_frag));
